Update Support for Spree::Address#name attribute
`#firstname` and `#lastname` are now deprecated in Solidus, and
they're going to be removed in 3.0 in favor of `#name`.

From 3.0 on, the firstname and the lastname we sent to Affirm
will need to be approximated to the first part of the name and
the remainder. For this task, we're using the support class
`Spree::Address::Name` provided by Solidus starting from v2.11.

Co-authored-by: spaghetticode <andrea@spaghetticode.it>